+++ src/Component.js
@@ -143,6 +143,8 @@ export default class Component {
     else if ('attachEvent' in obj) {
       obj.attachEvent(`on${type}`, func);
     }
+
+    return this;
   }
 
   /**
@@ -162,6 +164,8 @@ export default class Component {
     if (indexes.length) {
       _.pullAt(this.eventHandlers, indexes);
     }
+
+    return this;
   }
 
   /**
@@ -193,9 +197,8 @@ export default class Component {
    * @param container
    */
   appendTo(element, container) {
-    if (container) {
-      container.appendChild(element);
-    }
+    container?.appendChild(element);
+    return this;
   }
 
   /**
@@ -219,6 +222,8 @@ export default class Component {
         container.appendChild(element);
       }
     }
+
+    return this;
   }
 
   /**
@@ -236,6 +241,8 @@ export default class Component {
         console.warn(err);
       }
     }
+
+    return this;
   }
 
   /**
@@ -268,9 +275,7 @@ export default class Component {
    */
   appendChild(element, child) {
     if (Array.isArray(child)) {
-      child.forEach(oneChild => {
-        this.appendChild(element, oneChild);
-      });
+      child.forEach((oneChild) => this.appendChild(element, oneChild));
     }
     else if (child instanceof HTMLElement || child instanceof Text) {
       element.appendChild(child);
@@ -278,6 +283,8 @@ export default class Component {
     else if (child) {
       element.appendChild(this.text(child.toString()));
     }
+
+    return this;
   }
 
   /**
@@ -395,12 +402,15 @@ export default class Component {
    */
   addClass(element, className) {
     if (!element) {
-      return;
+      return this;
     }
+
     const classes = element.getAttribute('class');
-    if (!classes || classes.indexOf(className) === -1) {
+    if (!classes?.includes(className)) {
       element.setAttribute('class', `${classes} ${className}`);
     }
+
+    return this;
   }
 
   /**
@@ -413,13 +423,15 @@ export default class Component {
    */
   removeClass(element, className) {
     if (!element) {
-      return;
+      return this;
     }
-    let cls = element.getAttribute('class');
-    if (cls) {
-      cls = cls.replace(new RegExp(` ${className}`, 'g'), '');
-      element.setAttribute('class', cls);
+
+    const classes = element.getAttribute('class');
+    if (classes) {
+      element.setAttribute('class', classes.replace(new RegExp(` ${className}`, 'g'), ''));
     }
+
+    return this;
   }
 
   /**